+++ src/misc/filters/index.js
@@ -5,6 +5,7 @@ import * as Volume from './audio/Volume';
 import * as Loudnorm from './audio/Loudnorm';
 
 // Video Filter
+import * as Scale from './video/Scale';
 import * as Transpose from './video/Transpose';
 import * as HFlip from './video/HFlip';
 import * as VFlip from './video/VFlip';
@@ -51,6 +52,7 @@ audioRegistry.Register(Loudnorm);
 
 // Video Filters
 const videoRegistry = new Registry('video');
+videoRegistry.Register(Scale);
 videoRegistry.Register(Transpose);
 videoRegistry.Register(HFlip);
 videoRegistry.Register(VFlip);